feat: calculate cell geo locations on demand (#125)
This changes the way cell geo locations are calculated and stored in the cache table. Previously the process was run for every incoming message, which resulted in a lot of unused computation results.

This is now changed so that the UI can request cell geo locations through executing a state machine and continue to use the cache table. [It means a little extra work for the UI](https://github.com/bifravst/aws/pull/125/files#diff-15a3991c857e4b76a001ee53bfc92f97), but now the computations are only done if really needed by the UI.

In the future the workflow of querying the cache table and falling back to the step function could be combined into one lambda behind and HTTP API.

BREAKING CHANGE: the cell geolocation cache table is no longer automatically filled