feat(cdk): unify the way APIs are configured
Previously there were two ways to configure APIs
(one for UnwiredLabs and one for GitHub).

This is now unified and there is a CLI to set
these configuration settings as well.

In addition all SSM Parameters are now prefixed with the
stack name so multiple instances can exist in the same
account.

This is not marked as a breaking change because
the documentation was not yet published