test(unwiredlabs): set up a mock HTTP api for testing the integration
This sets up a mock HTTP API which stores
incoming requests in SQS for inspection and
can optionally respond with queued replies.

This HTTP API endpoint is then used in the
new e2e test that tests that the Unwired Labs
integration works.

This requires the set up of a separate stack
and in order to reuse the lambda packing utils
some refactoring of the code that handles packing
of lambdas was neccessary.

This is a preparation for #41, where we need to
query the Unwired Labs API differently based on
the network mode a device is which reports the
cell.