feat(iot): allow devices to subscribe to device topics
See #460

This allows devices to subscribe to any topic below their device ID prefix.

This will be used for transmitting A-GPS and P-GPS data.

BREAKING CHANGE: this will remove the existing policy from the thing
group. Re-attach the `${STACK_NAME}-thingPolicy-...` to the `${STACK_NAME}` thing group in the AWS IoT console.