feat: store web app configuration in SSM
This ensures that only the neccessary parameters are exposed
and that read access to CloudFormation is not neccessary for the deployment
of the web application. It also serves as an allowlist of parameters that
are exposed so that other CloudFormation outputs are not included by accident
in the web app's configuration.