feat: remove avatar storage
The avatars are more of a playful feature and not
really an essential feature to demonstrate so in
order to keep the codebase concise it is removed.

BREAKING CHANGE: this removes the avatar storage

When updating an existing installation, deploy
the `webapp` stack first, so the reference to
the avatar storage is removed.